+++ Sprocket/Objects/Cameras/Camera.h
@@ -13,7 +13,7 @@ public:
     virtual Maths::mat4 View() const = 0;
     virtual Maths::vec3 Position() const = 0;
 
-    virtual void OnUpdate(float timeDelta) {}
+    virtual void OnUpdate(double timeDelta) {}
     virtual void OnEvent(Event& event) {}
 };
